requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
boto3==1.34.40
botocore==1.34.40
certifi==2024.2.2
cffi==1.16.0
Django==5.0.1
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==7.0.0
gunicorn==21.2.0
inflection==0.5.1
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
minio==7.2.3
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pyflakes==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
